remove union struct, now 100%(?) c89 compliant
found out that you can only have int / long in bitfield in c89. this will probably be the last time i ever target c89 as it's beyond awful haha.

i've added "getters and setters" for each bit in the flags. i imagine this to be slightly slower, but oh well