add back non-c89 code, make c89 optional
while i haven't tested the speed yet, the c89 version produces 500 lines more assembly, and by looking at that assembly it's easy to see that it would be slower.

because of this, i have added back the non-c89 code which is enabled by default.

to use c89 code, define C89 when compiling.