implement decimal arithmetic mode
The implementation closely follows appendix A from
http://www.6502.org/tutorials/decimal_mode.html and reverts to the
original ADCSBC macro when decimal mode is disabled.

Decimal mode can be disabled, as necessary for 2a03 emulation, with a
flag in the mos6502 struct.

The changes are verified using a table, "arithmetic.bin", which for each
operation (adc/sbc), for each combination of carry and decimal flags and
for each possible A and operand value contains the expected A and flag
content.